Add `generate_annotations_json_file` function.
Dumps all HMM documents into a JSON file.
